Fix typos, improve data types, and enhance application logic
Updated interface names and data types for better precision.
Renamed data grid views for clarity and introduced new
controllers for handling applications and application types.
Enhanced filtering logic and updated project files to
reflect these changes, improving overall structure and
functionality.

+++ Core/Models/ApplicationType.cs
@@ -6,6 +6,6 @@ namespace Core.Models
     {
         public int ApplicationTypeId { get; set; }
         public string ApplicationTypeName { get; set; }
-        public float ApplicationTypeFees { get; set; }
+        public decimal ApplicationTypeFees { get; set; }
     }
 }
